% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{addBatchClusters}
\alias{addBatchClusters}
\title{Single-Cell Utilities Constructor}
\usage{
addBatchClusters(project, resolutions = 0.5, ...)
}
\arguments{
\item{project}{ArchR project with meta-data}

\item{resolutions}{vector}
}
\description{
Single-Cell Utilities Constructor
}
