% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DimensionReduction.R
\name{addPCA}
\alias{addPCA}
\title{addPCA
add PCA}
\usage{
addPCA(
  project,
  useMatrix = "GeneExpressionMatrix",
  assay = "data",
  nPCs = 50,
  name = "PCA",
  reduction.key = "PC_",
  seed = 1,
  approx = TRUE,
  verbose = FALSE,
  force = TRUE,
  ...
)
}
\value{
project
}
\description{
addPCA
add PCA
}
