% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processing.R
\name{Cellcycling}
\alias{Cellcycling}
\title{Single-Cell Processing}
\usage{
Cellcycling(project, force = T)
}
\arguments{
\item{project}{ArchR project}

\item{force=T, }{if TRUE overwrite, ELSE add a {name}_new}
}
\description{
Single-Cell Processing
}
