% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R, R/generics.R
\name{getCellCol.ArchRProject}
\alias{getCellCol.ArchRProject}
\alias{getCellCol.Seurat}
\alias{getCellCol}
\title{getCellCol function}
\usage{
\method{getCellCol}{ArchRProject}(object, ...)

\method{getCellCol}{Seurat}(object)

getCellCol(object, ...)
}
\arguments{
\item{object}{Seurat or ArchR object}
}
\value{
CellColDF

CellColDF

CellColDF
}
\description{
get cell column meta

get cell column meta
}
\examples{
getCellCol()
getCellCol()
}
