% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R, R/generics.R
\name{getDimRed.ArchRProject}
\alias{getDimRed.ArchRProject}
\alias{getDimRed.Seurat}
\alias{getDimRed}
\title{getDimRed function}
\usage{
\method{getDimRed}{ArchRProject}(object, reduction, ...)

\method{getDimRed}{Seurat}(object, reduction, ...)

getDimRed(object, ...)
}
\arguments{
\item{object}{Seurat or ArchR object}

\item{reduction}{reduction dataframe}
}
\value{
Dimension reduction

Dimension reduction

Dimension reduction
}
\description{
get dimension reduction

get dimension reduction
}
\examples{
getDimRed()
getDimRed()
}
