% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R, R/generics.R
\name{getMatrix.ArchRProject}
\alias{getMatrix.ArchRProject}
\alias{getMatrix.Seurat}
\alias{getMatrix}
\title{getMatrix function}
\usage{
\method{getMatrix}{ArchRProject}(object, matrix_name, assay = "counts", rowname = "meta", ...)

\method{getMatrix}{Seurat}(object, assay = "RNA", slot = "counts")

getMatrix(object, ...)
}
\arguments{
\item{object}{Seurat or ArchR object}

\item{matrix_name}{matrix name}

\item{assay}{assay name}

\item{rowname}{where is rowname from, meta or rowRanges}

\item{slot}{counts, data, scale.data}
}
\value{
Matrix

Matrix

Matrix
}
\description{
get Matrix

get Matrix
}
\examples{
getMatrix()
getMatrix()
}
