% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportFeatureMatrix.R
\name{import10xFeatureMatrix_}
\alias{import10xFeatureMatrix_}
\title{Import Feature Matrix from 10x Feature HDF5 file.}
\usage{
import10xFeatureMatrix_(
  input = NULL,
  input_type = NULL,
  names = NULL,
  strictMatch = TRUE,
  verbose = TRUE,
  featureType = "Gene Expression"
)
}
\arguments{
\item{input}{A character of paths to 10x feature hdf5 file(s) or 10X sparse matrix directory. These will traditionally have a suffix similar to "filtered_feature_bc_matrix.h5".}

\item{input_type}{A character of input types 10XH5 or 10X}

\item{names}{A character of sample names associated with each input file.}

\item{strictMatch}{Only relevant when multiple input files are used. A boolean that indictes whether rows (genes) that do not match perfectly in the matrices
should be removed (`strictMatch = TRUE`) or coerced (`strictMatch = FALSE`). CellRanger seems to occassionally use different ensembl ids for the same gene across
different samples. If you are comfortable tolerating such mismatches, you can coerce all matrices to fit together, in which case the gene metadata present in
the first listed sample will be applied to all matrices for that particular gene entry. Regardless of what value is used for `strictMatch`, this function
cannot tolerate mismatched gene names, only mismatched metadata for the same gene.}

\item{verbose}{Only relevant when multiple input files are used. A boolean that indicates whether messaging about mismatches should be verbose (`TRUE`) or minimal (`FALSE`)}

\item{featureType}{The name of the feature to extract from the 10x feature file.
See https://support.10xgenomics.com/single-cell-gene-expression/software/pipelines/latest/advanced/h5_matrices for more information.}
}
\description{
This function will import the feature matrix from a 10x feature hdf5 file.
}
