% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scProportion.R
\name{sc_utils}
\alias{sc_utils}
\title{Single-Cell Utilities Constructor}
\usage{
sc_utils(project)
}
\arguments{
\item{project}{ArchR project with meta-data}
}
\description{
Single-Cell Utilities Constructor
}
