% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R, R/generics.R
\name{setCellCol.ArchRProject}
\alias{setCellCol.ArchRProject}
\alias{setCellCol.Seurat}
\alias{setCellCol}
\title{setCellCol function}
\usage{
\method{setCellCol}{ArchRProject}(object, meta, ...)

\method{setCellCol}{Seurat}(object, meta)

setCellCol(object, ...)
}
\arguments{
\item{object}{Seurat or ArchR object}

\item{meta}{meta_data data.frame}
}
\value{
object

OBJ

OBJ
}
\description{
set cell column meta

set cell column meta
}
\examples{
setCellCol()
setCellCol()
}
