% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R, R/generics.R
\name{setMatrix.ArchRProject}
\alias{setMatrix.ArchRProject}
\alias{setMatrix.Seurat}
\alias{setMatrix}
\title{setMatrix function}
\usage{
\method{setMatrix}{ArchRProject}(
  object = NULL,
  mtx = NULL,
  matrix_name,
  force = F,
  matrix_property = "gene",
  entry_type = "double"
)

\method{setMatrix}{Seurat}(object, mtx = NULL, assay = NULL, force = F)

setMatrix(object, ...)
}
\arguments{
\item{object}{Seurat or ArchR object}

\item{mtx}{matrix}

\item{matrix_name}{matrix name}

\item{force}{force to overwrite}

\item{matrix_property}{property}

\item{entry_type}{double integer ...}

\item{assay}{assay name}
}
\value{
object

object

OBJ
}
\description{
set Matrix

set Matrix
}
\examples{
setMatrix()
setMatrix()
}
